backend_django/AI/requirements.txt
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.5
click==8.0.1
colorama==0.4.4
cycler==0.10.0
Django==2.2.24
django-cors-headers==3.5.0
django-seed==0.2.2
djangorestframework==3.12.4
easyocr==1.4.1
Faker==5.0.0
fire==0.4.0
gTTS==2.2.3
idna==3.2
imageio==2.9.0
importlib-metadata==4.8.1
joblib==1.0.1
kiwisolver==1.3.2
lmdb==1.2.1
matplotlib==3.4.3
natsort==7.1.1
networkx==2.6.3
nltk==3.6.2
numpy==1.21.2
opencv-python-headless==4.5.3.56
Pillow==8.2.0
pycodestyle==2.7.0
pyparsing==2.4.7
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
scikit-image==0.18.3
scipy==1.7.1
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
text-unidecode==1.3
tifffile==2021.8.30
toml==0.10.2
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.2
typing-extensions==3.10.0.2
urllib3==1.26.6
wincertstore==0.2
zipp==3.5.0
